+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Version 73
 *in development*
 
 * 🌟 Discover: add new episodes screen with year and language filters, similar to popular shows screen.
+* 🔧 Stream and purchase filter: sort providers by name.
 * 🔧 History: sort Trakt friends by latest activity first.
 * 🔨 Comments: do not display misleading "Could not modify data" message when commenting on a show.